implement timezone
- Timezone implemented

Going to restructure the app so that it creates a directory in which it can write saved settings to.

Currently it runs when compiled as a jar but errors out if you try to apply or reset the settings; the above change that I will make sometime next week will fix that.