update shortcut
The vbscript file now finds a file that ends with the .jar extension and uses that in the shortcut (if someone renames the file then this way it won't break.)

Also made the target path directly reference javaw so it doesn't have to get the java_home variable, which not all users may have.

Implemented dark theme for the error message background.