fix bugs, missing image notif
Fixed a bug in advanced settings, the program now displays an error if the image data did not load, is missing, or was moderated.

Updates will be more scarce henceforth, since school recently started up again for me; most of the features I plan to add will require a great deal of time and concentration, which, currently, I have to spend on school.